requirements.txt
# Core dependencies
streamlit==1.32.0
sentence-transformers==2.2.2
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
uvicorn

# Specific versions to ensure compatibility
huggingface-hub<0.19.0,>=0.18.0
torch>=1.6.0
transformers>=4.6.0,<5.0.0
numpy>=1.19.5
scikit-learn>=0.24.1
tqdm>=4.41.0